refactor(core): rename skeletons/ directory to builders/
